Add Page

+++ app/dashboard/thread.tsx
@@ -1,10 +1,15 @@
-import { Group, Stack, Text } from "@chakra-ui/react";
+import { IconButton, Stack } from "@chakra-ui/react";
 import type { Route } from "./+types/thread";
 import { prisma } from "~/prisma";
 import { redirect, useFetcher } from "react-router";
 import ChatBox from "./chat-box";
 import { getAuthUser } from "~/auth/middleware";
 import { createToken } from "~/jwt";
+import { Page } from "~/components/page";
+import { getThreadName } from "~/thread-util";
+import { useContext, useEffect, useState } from "react";
+import { AppContext } from "./context";
+import { TbCheck, TbMessage, TbTrash } from "react-icons/tb";
 
 export async function loader({ params, request }: Route.LoaderArgs) {
   const thread = await prisma.thread.findUnique({
@@ -30,24 +35,59 @@ export async function action({ params }: Route.ActionArgs) {
 
 export default function Thread({ loaderData }: Route.ComponentProps) {
   const deleteFetcher = useFetcher();
+  const { threadTitle } = useContext(AppContext);
+  const [deleteActive, setDeleteActive] = useState(false);
+
+  useEffect(() => {
+    if (deleteActive) {
+      setTimeout(() => {
+        setDeleteActive(false);
+      }, 3000);
+    }
+  }, [deleteActive]);
 
   function handleDelete() {
+    if (!deleteActive) {
+      setDeleteActive(true);
+      return;
+    }
     deleteFetcher.submit(null, {
       method: "delete",
     });
   }
 
   return (
-    <Stack p={8} h="full">
+    <Page
+      title={
+        threadTitle[loaderData.thread.id] ??
+        getThreadName(loaderData.thread.messages)
+      }
+      icon={<TbMessage />}
+      right={
+        <IconButton
+          size={"xs"}
+          variant={"subtle"}
+          onClick={handleDelete}
+          colorPalette={
+            deleteActive || deleteFetcher.state !== "idle" ? "red" : undefined
+          }
+          disabled={deleteFetcher.state !== "idle"}
+        >
+          {deleteActive || deleteFetcher.state !== "idle" ? (
+            <TbCheck />
+          ) : (
+            <TbTrash />
+          )}
+        </IconButton>
+      }
+    >
       <Stack maxW={"700px"}>
         <ChatBox
           token={loaderData.token}
           thread={loaderData.thread}
           key={loaderData.thread.id}
-          deleting={deleteFetcher.state !== "idle"}
-          onDelete={handleDelete}
         />
       </Stack>
-    </Stack>
+    </Page>
   );
 }
